add entities modificated Creation Account Controller Created repository to find songs by title, album, genre Add methods to add Song Add set Modified method and inserted an exception if the registration fields such as title and artist are not inserted Modified Song's Controller to add song. Add method to delete song to @Service Add controller to delete song Processing method to update edited method update, fixed error Add message to error in methods to update song and i try to put exception in addSong Add boolean to see if song is public or not. Add input to method addSong and try-catch-finally to close Scanner

+++ src/main/java/com/dhmusic/DHMusic/services/SongService.java
@@ -5,6 +5,8 @@ import com.dhmusic.DHMusic.repositories.content.repositories.SongRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Scanner;
+
 @Service
 public class SongService {
 
@@ -16,7 +18,18 @@ public class SongService {
     }
 
     public Song addSong(Song song){
-        Song existSong = (Song) songRepository.findSongByTitle(song.getTitle()); //TODO
+        Scanner scanner = new Scanner(System.in);
+        try {
+            System.out.println("Enter song title:");
+            String title = scanner.nextLine();
+
+            System.out.println("Enter artist name:");
+            String artist = scanner.nextLine();
+
+            System.out.println("Enter album name:");
+            String album = scanner.nextLine();
+
+        Song existSong = (Song) songRepository.findSongByTitle(song.getTitle());
               //  && songRepository.findSongByArtist(song.getArtistOfSong());
         if (existSong != null){
             throw new RuntimeException("Song exist!");
@@ -24,16 +37,28 @@ public class SongService {
         if (song.getTitle() == null || song.getArtistOfSong() == null ) {
             throw new IllegalArgumentException("Mistake! Required fields are missing");
         }
+        Song newSong = new Song();
+
         return songRepository.save(song);
-    }
+    }catch (Exception e){
+            System.out.println("An error has occurred:"+ e.getMessage());
+            return null;
+        }finally {
+            scanner.close();
+        }
+        }
 
     public void deleteSong (Song song){
-        // bisogna mettere eccezione se utente è admin?
+        Song existSong = (Song) songRepository.findSongByTitle(song.getTitle()); //mettere indice
+        if (existSong == null){
+            throw new RuntimeException("Song not exist!");
+        }
+        //if() // TODO se la canzone è pubblica e l'utente è admin, cancella
         songRepository.delete(song);
     }
 
     public Song updateSong(Song song){
-        Song existSong = (Song) songRepository.findSongById(song.getIdSong());
+        Song existSong = (Song) songRepository.findSongById(song.getIdSong()); // id string o long?
         if (existSong == null){
             throw new RuntimeException("Song not exist");
             //return null;
